move system to its own mod
