+++ reswmm-cli/src/main.rs
@@ -1,5 +1,5 @@
 use reswmm_core::{
-    units::{qnty::Qnty, si::Length},
+    units::{qnty::Qnty, system::si::Length},
     xsection,
 };
