+++ reswmm-core/src/lib.rs
@@ -4,7 +4,7 @@ pub mod xsection;
 mod tests {
     use super::*;
     use xsection::{Rectangle, XSection};
-    use units::{qnty::Qnty, si::{Length, Area}};
+    use units::{qnty::Qnty, system::si::{Length, Area}};
     #[test]
     fn xsection() {
         let xs = Rectangle::new(Qnty::<Length>::new(2.0));
